requirements-dev.txt
# Course Record Updater - Development Dependencies
# Quality assurance and development tools

# Code formatting and style
black>=23.0.0              # Python code formatter
isort>=5.12.0              # Import statement organizer

# Linting and static analysis
flake8>=6.0.0              # Style guide enforcement
pylint>=2.17.0             # Comprehensive static analysis
mypy>=1.5.0                # Static type checker

# Testing and coverage
pytest>=7.4.0              # Testing framework
pytest-cov>=4.1.0          # Coverage reporting for pytest
pytest-mock>=3.11.0        # Mock utilities for pytest
pytest-asyncio>=0.21.0     # Async testing support
pytest-xdist>=3.6.1        # Parallel test execution (35% faster)

# Security scanning
bandit>=1.7.5              # Security linter for Python
safety>=2.3.0              # Dependency vulnerability scanner
detect-secrets>=1.4.0        # Secret detection in code

# Code quality and complexity
radon>=6.0.1               # Code complexity analysis
xenon>=0.9.0               # Code complexity monitoring
diff-cover>=10.0.0         # Coverage analysis on new/modified code (PR quality gate)

# Development utilities
pre-commit>=3.3.0          # Git hooks for quality checks
tox>=4.6.0                 # Testing across multiple Python versions

# Security patch minimums (safety scan)
filelock>=3.20.1
authlib>=1.6.5
requests>=2.32.4

# Documentation
sphinx>=7.1.0              # Documentation generator
sphinx-rtd-theme>=1.3.0    # Read the Docs theme

# Type stubs for better type checking
types-requests>=2.31.0     # Type stubs for requests library
types-python-dateutil>=2.8.0  # Type stubs for dateutil

# Flask development tools
flask-testing>=0.8.1       # Flask testing utilities
Flask-WTF>=1.2.1           # CSRF protection and form handling (needed for tests)

# Data processing (needed for Excel import tests)
pandas>=2.3.2               # Data manipulation and analysis
openpyxl>=3.1.0             # Excel file support for pandas

# Browser automation for E2E testing
playwright>=1.55.0          # Modern browser automation (faster, more reliable than Selenium)
pytest-playwright>=0.7.1    # Pytest plugin for Playwright integration
pytest-base-url>=2.1.0      # Base URL fixtures for E2E tests
selenium>=4.15.0            # Legacy browser automation (still used by some integration tests)
defusedxml>=0.7.1
